+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Change methods to return shared_key and allow decryption from it [#7]
 - Change encryption gadget to include bad encryption check [#8]
+- Change encryption methods to allow custom generators [#11]
 
 ## [0.2.0] - 2025-02-06
 
@@ -27,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add initial implementation [#1]
 
 <!-- ISSUES -->
+[#11]: https://github.com/dusk-network/jubjub-elgamal/issues/11
 [#8]: https://github.com/dusk-network/jubjub-elgamal/issues/8
 [#7]: https://github.com/dusk-network/jubjub-elgamal/issues/7
 [#4]: https://github.com/dusk-network/jubjub-elgamal/issues/4